feat: Add basic Benford analysis tools
User can explore their own datasets with the benford object to see if
their data conforms to Benford's law of anomalous numbers. Capabilities
for exploring various radix, starting digits, and consecutive digits are
built into the analysis. Comes with confidence bounds for rejecting the
null hypothesis, so that the user can quantify if their data conforms to
Benford's law or not.